Create new package jsonschema and spec
